+++ crates/parol/src/generators/grammar_config.rs
@@ -21,8 +21,11 @@ static DEFAULT_GRAMMAR_CONFIG: Lazy<GrammarConfig> = Lazy::new(GrammarConfig::de
 /// the grammar, user type aliases, maximum lookahead size and a list of scanner configurations.
 /// Can later be amended with further information or pragmas that control lexer/parser generation.
 ///
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default)]
 pub struct GrammarConfig {
+    /// The non-terminals of the original context free grammar from the user's grammar description
+    /// before any checks and transformation has been conducted.
+    pub non_terminals: Vec<String>,
     ///
     /// The actual context free grammar.
     /// It should be checked and left-factored here.
@@ -71,12 +74,15 @@ impl GrammarConfig {
     pub fn new(cfg: Cfg, lookahead_size: usize) -> Self {
         Self {
             cfg,
-            title: None,
-            comment: None,
-            user_type_defs: Vec::new(),
-            scanner_configurations: Vec::new(),
             lookahead_size,
+            ..Default::default()
         }
+    }
+
+    /// Sets the non-terminals
+    pub fn with_non_terminals(mut self, non_terminals: Vec<String>) -> Self {
+        self.non_terminals = non_terminals;
+        self
     }
 
     /// Sets an optional title
@@ -108,7 +114,7 @@ impl GrammarConfig {
         self.lookahead_size = k;
     }
 
-    /// Sets the cfg member
+    /// Updates the cfg member after the grammar has been checked and transformed
     pub fn update_cfg(&mut self, cfg: Cfg) {
         self.cfg = cfg;
     }
@@ -177,19 +183,6 @@ impl GrammarConfig {
     /// Generates a dummy user_type_resolver function that can be used in Pr::format
     pub fn dummy_user_type_resolver() -> FnUserTypeResolver {
         Box::new(|_u: &str| None)
-    }
-}
-
-impl Default for GrammarConfig {
-    fn default() -> Self {
-        Self {
-            cfg: Cfg::default(),
-            title: None,
-            comment: None,
-            user_type_defs: Vec::new(),
-            scanner_configurations: vec![ScannerConfig::default()], // There must always be a default scanner
-            lookahead_size: 0,
-        }
     }
 }